requirements.txt
accelerate==1.0.1
asttokens==3.0.0
attrs==25.3.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
beautifulsoup4==4.13.5
bitsandbytes==0.42.0
bleach==6.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.1.1
cycler==0.12.1
decorator==4.4.2
defusedxml==0.7.1
docopt==0.6.2
easydict==1.13
eduvpn-client==4.1.1
eduvpn-common==1.1.0
einops==0.8.1
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.16.1
fonttools==4.57.0
fsspec==2024.10.0
huggingface-hub==0.26.5
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1620912944151/work/dist/ipykernel-5.5.5-py3-none-any.whl
ipython==8.12.3
ipython_genutils @ file:///home/conda/feedstock_root/build_artifacts/ipython_genutils_1716278396992/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1669134318875/work
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1633454794268/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1748333051527/work
jupyterlab_pygments==0.3.0
kiwisolver==1.4.7
lazy_loader==0.4
lightning-utilities==0.11.9
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mistune==3.1.4
monai==1.3.2
mpmath==1.3.0
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
opencv-python==4.12.0.88
packaging==24.2
pandas==2.0.3
pandocfilters==1.5.1
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.4.0
pipreqs==0.5.0
pkgutil_resolve_name==1.3.10
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1715777629804/work
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1717574682739/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure_eval==0.2.3
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
pyparsing==3.1.4
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytorch-metric-learning==2.9.0
pytz==2025.2
PyWavelets==1.4.1
PyYAML @ file:///croot/pyyaml_1728657952215/work
pyzmq==19.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.4
rpds-py==0.20.1
safetensors==0.4.5
sarrarp50-toolkit @ git+https://github.com/surgical-vision/SAR_RARP50-evaluation@519280bf52034b0931a63688adee7ea16bcc42a2
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
soupsieve==2.7
stack-data==0.6.3
sympy==1.13.3
threadpoolctl==3.5.0
tifffile==2023.7.10
tinycss2==1.2.1
tokenizers==0.20.3
torch==2.4.1
torchmetrics==1.5.2
torchvision==0.19.1
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827257044/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
transformers==4.46.3
triton==3.0.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
webencodings==0.5.1
yarg==0.1.9
zipp==3.20.2
